+++ resources/themes/pterodactyl/admin/servers/view/startup.blade.php
@@ -109,6 +109,18 @@
                     </div>
                 </div>
             </div>
+            <div class="box">
+                <div class="box-header with-border">
+                    <h3 class="box-title">Docker Container Configuration</h3>
+                </div>
+                <div class="box-body">
+                    <div class="form-group">
+                        <label for="pDockerImage" class="control-label">Image</label>
+                        <input type="text" name="docker_image" id="pDockerImage" value="{{ $server->image }}" class="form-control" />
+                        <p class="text-muted small">The Docker image to use for this server. The default image for the selected egg is <code id="setDefaultImage"></code>.</p>
+                    </div>
+                </div>
+            </div>
         </div>
         <div class="col-md-6">
             <div class="row" id="appendVariablesTo"></div>
@@ -122,24 +134,9 @@
     {!! Theme::js('vendor/lodash/lodash.js') !!}
     <script>
     $(document).ready(function () {
-        $('#pNestId').select2({
-            placeholder: 'Select a Nest',
-        }).change();
-        $('#pEggId').select2({
-            placeholder: 'Select a Nest Egg',
-        });
-        $('#pPackId').select2({
-            placeholder: 'Select a Service Pack',
-        });
-
-        $('input[data-action="match-regex"]').on('keyup', function (event) {
-            if (! $(this).data('regex')) return;
-
-            var input = $(this).val();
-            var regex = new RegExp($(this).data('regex').replace(/^\/|\/$/g, ''));
-
-            $(this).parent().parent().removeClass('has-success has-error').addClass((! regex.test(input)) ? 'has-error' : 'has-success');
-        });
+        $('#pNestId').select2({placeholder: 'Select a Nest'}).change();
+        $('#pEggId').select2({placeholder: 'Select a Nest Egg'});
+        $('#pPackId').select2({placeholder: 'Select a Service Pack'});
     });
     </script>
     <script>
@@ -158,7 +155,11 @@
             var parentChain = _.get(Pterodactyl.nests, $('#pNestId').val(), null);
             var objectChain = _.get(parentChain, 'eggs.' + $(this).val(), null);
 
-            $('#pDefaultContainer').val(_.get(objectChain, 'docker_image', 'not defined!'));
+            $('#setDefaultImage').html(_.get(objectChain, 'docker_image', 'undefined'));
+            $('#pDockerImage').val(_.get(objectChain, 'docker_image', 'undefined'));
+            if (objectChain.id === parseInt('{{ $server->egg_id }}')) {
+                $('#pDockerImage').val('{{ $server->image }}');
+            }
 
             if (!_.get(objectChain, 'startup', false)) {
                 $('#pDefaultStartupCommand').val(_.get(parentChain, 'startup', 'ERROR: Startup Not Defined!'));
@@ -192,7 +193,7 @@
                                 <h3 class="box-title">' + isRequired + item.name + '</h3> \
                             </div> \
                             <div class="box-body"> \
-                                <input data-action="match-regex" name="environment[' + item.env_variable + ']" class="form-control" type="text" value="' + setValue + '" /> \
+                                <input name="environment[' + item.env_variable + ']" class="form-control" type="text" id="egg_variable_' + item.env_variable + '" /> \
                                 <p class="no-margin small text-muted">' + item.description + '</p> \
                             </div> \
                             <div class="box-footer"> \
@@ -202,6 +203,7 @@
                         </div> \
                     </div>';
                 $('#appendVariablesTo').append(dataAppend);
+                $('#appendVariablesTo').find('#egg_variable_' + item.env_variable).val(setValue);
             });
         });
     </script>